Starting integration into pipeline and environment
bindsnet/pipeline

- Pipeline is being split into multiple pieces
  - General philosophy being that we provide useful components to
  inherit from and concrete examples for the datasets and environments
  that we plan on supporting.

base_pipeline.py
- BasePipeline mainly responsible for high level functions, when to
save, when to plot, etc.

dataloader_pipeline.py
- DataLoaderPipeline sets up the training loop for a specified number of
epochs and manages the DataLoader object from pytorch. Subclasses need
to implement a small subset of functions, but this ultimately allows for
a very flexible end implementation.

- TorchVisionDatasetPipeline is a specific implementation that follows
the dataset setup from the bindsnet/datasets/torchvision_wrapper.py and
is provided as an example of how to subclass DataLoaderPipeline.

bindsnet/datasets

torchvision_wrapper.py - switch to a dictionary for the batch instead of
a tuple. This provides the ability to add more information without
breaking argument order later down the line.

spike_encoders.py - fix the usage of dt