Expand on BasePipeline subclasses
The previous Pipeline object has been maintained and migrated to
RLPipeline as there are a number of components in there that seem to be
specific to RL.

BasePipeline provides some more convinence functions that were found
originally in Pipeline. Making an attempt to not keep to many attributes
inside of BasePipeline.

Update of some examples to work on the implementations of the
BasePipeline subclasses. Provide an example of a custom pipeline's
implementation inside of mnist/minimal_reservoir.